refactor(pcblib): split mod.rs read/write paths into read_io/write_io
pcblib/mod.rs was ~2.5k lines mixing the container/public API, the read path, and the write path. Move the read-orchestration methods (the public 'read' entry + parse helpers) into pcblib/read_io.rs and the write-orchestration methods (the public 'write' entry + encode/metadata/3D-model-prep helpers) into pcblib/write_io.rs, each as a second 'impl PcbLib' — methods still call each other via self and private fields stay accessible (descendant modules). mod.rs keeps the struct definitions, entry points (new/open/save), CRUD, 3D-model accessors, and tests. mod.rs: 2494 -> 1449 lines. Behaviour-neutral; fmt/clippy clean, all tests pass. Addresses #105 (the largest module).
